services/graphql-service/requirements.txt
fastapi==0.109.0
strawberry-graphql[fastapi]==0.215.0

